flask/requirements.txt
flask
pandas
scikit-learn
mlflow
python-dotenv
boto3




